Updated Bert Qa Android to use Task Library
