+++ lite/examples/bert_qa/android/.gitignore
@@ -1,0 +1,1 @@
+/app/src/main/assets/
