+++ lite/examples/bert_qa/android/README.md
@@ -1,13 +1,28 @@
 # TensorFlow Lite BERT QA Android Example Application
 
+
+<img src="https://user-images.githubusercontent.com/67560900/122643946-37d0d380-d130-11eb-8e7c-f467b90cb0dd.mp4" width="400px" alt="Video">
+
 ## Overview
 
-This is an end-to-end example of BERT Question & Answer application built with
-TensorFlow 2.0, and tested on SQuAD dataset. The demo app provides 48 passages
-from the dataset for users to choose from, and gives 5 most possible answers
-corresponding to the input passage and query.
+This is an end-to-end example of [BERT] Question & Answer application built with
+TensorFlow 2.0, and tested on [SQuAD] dataset version 1.1. The demo app provides
+48 passages from the dataset for users to choose from, and gives 5 most possible
+answers corresponding to the input passage and query.
 
 These instructions walk you through running the demo on an Android device.
+
+### Model used
+
+[BERT], or Bidirectional Encoder Representations from Transformers, is a method
+of pre-training language representations which obtains state-of-the-art results
+on a wide array of Natural Language Processing tasks.
+
+This app uses [MobileBERT](https://tfhub.dev/tensorflow/tfjs-model/mobilebert/1 "MobileBERT"), a compressed version of [BERT] that runs 4x faster and
+has 4x smaller model size.
+
+For more information, refer to the [BERT github page][BERT].
+
 
 ## Build the demo using Android Studio
 
@@ -19,7 +34,7 @@ These instructions walk you through running the demo on an Android device.
 
 *   Android Studio 3.2 or later.
     - Gradle 4.6 or higher.
-    - SDK Build Tools 28.0.3 or higher.
+    - SDK Build Tools 29.0.2 or higher.
 
 *   You need an Android device or Android emulator and Android development
     environment with minimum API 15.
@@ -41,7 +56,7 @@ These instructions walk you through running the demo on an Android device.
 ### Running
 
 *   You need to have an Android device plugged in with developer options enabled
-    at this point. See [here](https://developer.android.com/studio/run/device)
+    at this point. See [here](https://developer.android.com/studio/run/device "Download Link")
     for more details on setting up developer devices.
 
 *   If you already have Android emulator installed in Android Studio, select a
@@ -66,3 +81,6 @@ cd lite/examples/bert_qa/android   # Folder for Android app.
 ```
 adb install app/build/outputs/apk/debug/app-debug.apk
 ```
+
+[BERT]: https://github.com/google-research/bert "Bert"
+[SQuAD]: https://rajpurkar.github.io/SQuAD-explorer/ "SQuAD"
