+++ lite/examples/bert_qa/android/app/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'com.android.application'
 
 android {
     compileSdkVersion 29
-    buildToolsVersion "29.0.0"
+    buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "org.tensorflow.lite.examples.bertapp"
         minSdkVersion 26
@@ -52,14 +52,10 @@ dependencies {
     implementation 'com.google.android.material:material:1.0.0'
 
     implementation 'com.google.code.gson:gson:2.8.5'
-    implementation 'com.google.guava:guava:28.1-android'
-    implementation ('org.tensorflow:tensorflow-lite:2.4.0')
-    implementation 'org.tensorflow:tensorflow-lite-metadata:0.1.0'
 
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'androidx.test:core:1.2.0'
-    testImplementation 'com.google.truth:truth:1.0'
-    testImplementation 'org.robolectric:robolectric:4.3.1'
+    //Task Text Library dependency
+    implementation 'org.tensorflow:tensorflow-lite-task-text:0.2.0'
+
     androidTestImplementation 'androidx.test:runner:1.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 }
